+++ src/app/chatbot-design-studio/cds-dashboard/cds-dashboard.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 
 
@@ -14,6 +15,7 @@ import { CdsPopupComponent } from './utils/cds-popup/cds-popup.component';
 import { ChangelogComponent } from 'src/app/modals/changelog/changelog.component';
 import { CdsModalActivateBotComponent } from 'src/app/modals/cds-modal-activate-bot/cds-modal-activate-bot.component';
 import { CdsPublishOnCommunityModalComponent } from 'src/app/modals/cds-publish-on-community-modal/cds-publish-on-community-modal.component';
+import { CdsAgentGeneratorComponent } from 'src/app/modals/cds-agent-generator/cds-agent-generator.component';
 import { WsChatbotService } from 'src/app/services/websocket/ws-chatbot.service';
 
 const routes: Routes = [
@@ -80,10 +82,12 @@ const routes: Routes = [
     //MODALS
     CdsModalActivateBotComponent,
     CdsPublishOnCommunityModalComponent,
+    CdsAgentGeneratorComponent,
 
   ],
   imports: [
     CommonModule,
+    FormsModule,
     RouterModule.forChild(routes),
     MaterialModule,
     TranslateModule,